+++ app/src/main/java/com/qihuan/photowidget/App.kt
@@ -3,6 +3,7 @@ package com.qihuan.photowidget
 import android.app.Application
 import com.qihuan.photowidget.core.database.di.dbModule
 import com.qihuan.photowidget.di.appModule
+import com.qihuan.photowidget.feature.about.di.aboutModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -26,7 +27,11 @@ class App : Application() {
         startKoin {
             androidLogger()
             androidContext(this@App)
-            modules(appModule, dbModule)
+            modules(
+                appModule,
+                dbModule,
+                aboutModule,
+            )
         }
     }
 }
