+++ app/src/main/java/com/qihuan/photowidget/di/AppModule.kt
@@ -1,7 +1,9 @@
 package com.qihuan.photowidget.di
 
+import com.qihuan.photowidget.AppInfoImpl
+import com.qihuan.photowidget.core.common.AppInfo
 import org.koin.dsl.module
 
 val appModule = module {
-
+    single<AppInfo> { AppInfoImpl(get()) }
 }
