+++ app/src/main/res/values/strings.xml
@@ -34,7 +34,6 @@
     <string name="click_to_refresh">点击刷新</string>
     <string name="add_widget_link">添加微件链接</string>
     <string name="delete_widget_link">删除微件链接</string>
-    <string name="app_icon">应用图标</string>
     <string name="delete_photo">删除相片</string>
     <string name="widget_photo">微件相片</string>
     <string name="wallpaper">壁纸</string>
@@ -51,24 +50,6 @@
     <string name="saving">保存中…</string>
     <string name="add_widget">添加微件</string>
     <string name="edit_widget">编辑微件</string>
-    <string name="about_version">版本：%s</string>
-    <string name="about_title_introduction">介绍</string>
-    <string name="about_introduction">一款相片小组件应用，可添加多张相片，可以打开应用、链接，支持图片裁剪、组件圆角以及边距调整。</string>
-    <string name="about_title_developer">开发者</string>
-    <string name="about_developer_name">齐欢</string>
-    <string name="about_developer_github">https://github.com/qihuan92</string>
-    <string name="about_developer_coolapk_profile_url">http://www.coolapk.com/u/3588942</string>
-    <string name="about_app_coolapk_app_url">https://www.coolapk.com/apk/com.qihuan.photowidget</string>
-    <string name="about_title_open_source_licenses">开源协议</string>
-    <string name="about">关于</string>
-    <string name="about_title_open_source_url">项目地址</string>
-    <string name="about_open_source_url">https://github.com/qihuan92/PhotoWidget</string>
-    <string name="about_title_link_coolapk">酷安</string>
-    <string name="about_title_link_github">GitHub</string>
-    <string name="about_title_link_weibo">微博</string>
-    <string name="about_title_link_telegram">Telegram</string>
-    <string name="about_developer_weibo">https://weibo.com/qihuan86</string>
-    <string name="about_developer_telegram">https://t.me/qihuan92</string>
     <string name="conform_delete_photo_link">是否删除相片链接</string>
     <string name="conform_delete_photo">是否删除该图片</string>
     <string name="tip_add_widget_title">如何添加相片微件</string>
@@ -78,7 +59,6 @@
     <string name="ignore_battery_optimizations_setting">设置</string>
     <string name="ignore_tip">不再提示</string>
     <string name="installed_app_page_title">选择应用</string>
-    <string name="about_developer_introduction">https://qihuan92.github.io</string>
     <string name="force_refresh_widget">强制刷新</string>
     <string name="alert_msg_force_refresh_widget">如果您遇到相片微件消失等显示异常的问题，可通过此功能刷新所有微件</string>
     <string name="alert_title_force_refresh_widget">刷新微件</string>
@@ -112,7 +92,6 @@
     <string name="default_radius">默认圆角</string>
     <string name="radius_unit_description">可点击条目选择圆角单位</string>
     <string name="default_scale_type">默认相片位置</string>
-    <string name="privacy_policy">隐私政策</string>
     <string name="widget_background">微件背景</string>
     <string name="add_widget_frame">添加微件相框</string>
     <string name="delete_widget_frame">删除微件相框</string>
