+++ core/common/src/main/java/com/qihuan/photowidget/core/common/AppInfo.kt
@@ -1,0 +1,13 @@
+package com.qihuan.photowidget.core.common
+
+/**
+ * 应用信息
+ *
+ * @author Qi
+ * @since 2022/10/3
+ */
+interface AppInfo {
+    val versionName: String
+    val versionCode: Int
+    val appName: String
+}
