+++ core/common/src/main/java/com/qihuan/photowidget/core/common/navigation/AboutNavigation.kt
@@ -1,0 +1,12 @@
+package com.qihuan.photowidget.core.common.navigation
+
+/**
+ * 关于模块导航
+ *
+ * @author Qi
+ * @since 2022/10/3
+ */
+object AboutNavigation {
+    const val PATH_ABOUT = "/about/main"
+    const val PATH_PRIVACY = "/about/privacy"
+}
