+++ core/common/src/main/res/drawable/ic_outline_info_24.xml
