+++ core/common/src/main/res/drawable/ic_round_arrow_back_24.xml
