+++ core/common/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="loading">加载中…</string>
     <string name="option_icon">选项图标</string>
     <string name="item_selection_dialog_title">请选择</string>
+    <string name="app_icon">应用图标</string>
 </resources>
