+++ core/model/src/main/java/com/qihuan/photowidget/core/model/Constants.kt
@@ -21,12 +21,6 @@ object BroadcastAction {
     const val APPWIDGET_DELETED = "com.qihuan.photowidget.APPWIDGET_DELETED"
 }
 
-object License {
-    const val MIT = "MIT License"
-    const val APACHE_2 = "Apache Software License 2.0"
-    const val GPL_V3 = "GNU general public license Version 3"
-}
-
 object FileExtension {
     const val JPEG = "jpg"
     const val PNG = "png"
