+++ feature/about/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+    <application>
+        <activity android:name=".activity.AboutActivity" />
+
+        <activity android:name=".activity.PrivacyPolicyActivity" />
+    </application>
+</manifest>
