+++ feature/about/src/main/java/com/qihuan/photowidget/feature/about/module/License.kt
@@ -1,0 +1,7 @@
+package com.qihuan.photowidget.feature.about.module
+
+object License {
+    const val MIT = "MIT License"
+    const val APACHE_2 = "Apache Software License 2.0"
+    const val GPL_V3 = "GNU general public license Version 3"
+}
