+++ feature/about/src/main/java/com/qihuan/photowidget/feature/about/module/LicenseItem.kt
@@ -1,0 +1,8 @@
+package com.qihuan.photowidget.feature.about.module
+
+data class LicenseItem(
+    val name: String,
+    val author: String,
+    val type: String,
+    val url: String,
+)
