+++ feature/about/src/main/res/drawable/ic_coolapk.xml
