+++ feature/about/src/main/res/drawable/ic_github.xml
