+++ feature/about/src/main/res/drawable/ic_launcher_background.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:angle="-45"
+        android:endColor="@color/launcher_icon_background_end_color"
+        android:startColor="@color/launcher_icon_background_start_color"
+        android:type="linear" />
+</shape>
