+++ feature/about/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="108dp"
+    android:height="108dp"
+    android:viewportWidth="108"
+    android:viewportHeight="108">
+
+    <group
+        android:scaleX="2.200039"
+        android:scaleY="2.200039"
+        android:translateX="27.599531"
+        android:translateY="27.599531">
+
+        <path
+            android:fillColor="@android:color/white"
+            android:pathData="M21,19V5c0,-1.1 -0.9,-2 -2,-2H5c-1.1,0 -2,0.9 -2,2v14c0,1.1 0.9,2 2,2h14c1.1,0 2,-0.9 2,-2zM8.5,13.5l2.5,3.01L14.5,12l4.5,6H5l3.5,-4.5z" />
+
+        <path
+            android:name="dot_path"
+            android:fillColor="@color/launcher_icon_sun_color"
+            android:pathData="M 6.5,4 m 0,5
+            a 2,2 0 1,0 4,0
+            a 2,2 0 1,0 -4,0" />
+    </group>
+</vector>
