+++ feature/about/src/main/res/drawable/ic_telegram.xml
