+++ feature/about/src/main/res/drawable/ic_weibo.xml
