+++ feature/about/src/main/res/layout/activity_privacy_policy.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".about.PrivacyPolicyActivity">
+    tools:context=".activity.PrivacyPolicyActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
