+++ feature/about/src/main/res/values/strings.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="privacy_policy">隐私政策</string>
+    <string name="about">关于</string>
+    <string name="about_title_introduction">介绍</string>
+    <string name="app_name" translatable="false">相片微件</string>
+    <string name="about_introduction">一款相片小组件应用，可添加多张相片，可以打开应用、链接，支持图片裁剪、组件圆角以及边距调整。</string>
+    <string name="about_title_developer">开发者</string>
+    <string name="about_developer_name">齐欢</string>
+    <string name="about_developer_introduction">https://qihuan92.github.io</string>
+    <string name="about_version">版本：%s</string>
+    <string name="about_developer_github">https://github.com/qihuan92</string>
+    <string name="about_developer_coolapk_profile_url">http://www.coolapk.com/u/3588942</string>
+    <string name="about_app_coolapk_app_url">https://www.coolapk.com/apk/com.qihuan.photowidget</string>
+    <string name="about_title_open_source_licenses">开源协议</string>
+    <string name="about_title_open_source_url">项目地址</string>
+    <string name="about_open_source_url">https://github.com/qihuan92/PhotoWidget</string>
+    <string name="about_title_link_coolapk">酷安</string>
+    <string name="about_title_link_github">GitHub</string>
+    <string name="about_title_link_weibo">微博</string>
+    <string name="about_title_link_telegram">Telegram</string>
+    <string name="about_developer_weibo">https://weibo.com/qihuan86</string>
+    <string name="about_developer_telegram">https://t.me/qihuan92</string>
+</resources>
