+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include(":core:model")
 include(":core:database")
 include(":core:analysis")
 include(":core:common")
+include(":feature:about")
